# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API del sitio",
        "description": "API para el manejo de datos del sitio",
        "contact": {
            "name": "Israel David Villaroel Moreno",
            "email": "israeldavidvm@gmail.com"
        },
        "license": {
            "name": "Licencia abierta",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "List all products",
                "description": "Retrieves a paginated list of all products in the system.",
                "operationId": "1ce27f90dd8c39cea9520bf45d58c7c2",
                "responses": {
                    "200": {
                        "description": "A paginated list of products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://localhost/api/v1/products?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://localhost/api/v1/products?page=1"
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": "null"
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "http://localhost/api/v1/products"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": "null"
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create a new product",
                "description": "Adds a new product to the database with its details.",
                "operationId": "acaa0d3fdc74fd4b534e70fc016cf7f1",
                "requestBody": {
                    "description": "Product data to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "description": "Product Schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "example": "Pc Gamer"
                    },
                    "description": {
                        "type": "string",
                        "example": "The best pc Gamer"
                    },
                    "price": {
                        "type": "integer",
                        "example": 1000
                    },
                    "currency_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "tax_cost": {
                        "type": "integer",
                        "example": 100,
                        "nullable": true
                    },
                    "manufacturing_cost": {
                        "type": "integer",
                        "example": 500,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-14T12:34:56Z",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-14T12:34:56Z",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>) (Ejemplo: Bearer 2|Cfz3yDjKqUh55AUI6I9nQQv6MEHsEqQvJToMDnJ7e7c8478a)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}